requirements.txt
numba==0.57.1
numpy==1.22.0
matplotlib==3.5.2
scipy==1.11.1
pyyaml==6.0.1
casadi
pandas
opencv-python
torch


